refactor(mentee-page): change directories
