+++ projects/common-components/src/lib/lesson/lesson.component.scss
