+++ src/app/app-routing.module.ts
@@ -3,7 +3,7 @@ import { Routes, RouterModule } from "@angular/router";
 import { SignInComponent } from "../../projects/core/src/lib/authorization/components/sign-in/sign-in.component";
 import { AuthGuard } from "projects/core/src/lib/authorization/guards/auth/auth.guard";
 import { RoleSelectComponent } from "../../projects/core/src/lib/authorization/components/role-select/role-select.component";
-import { MenteePageComponent } from "projects/core/src/lib/mentee-page/mentee-page.component";
+import { MenteePageComponent } from "projects/core/src/lib/components/mentee-page/mentee-page.component";
 // route guards
 
 const routes: Routes = [
